% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflowr-package.R
\docType{package}
\name{cashflowr-package}
\alias{cashflowr}
\alias{cashflowr-package}
\title{cashflowr: Manage Your Cashflow With Ease}
\description{
Supply a budget and let cashflowr take care of the rest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbin/cashflowr}
  \item Report bugs at \url{https://github.com/lddurbin/cashflowr/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{l.d.durbin@gmail.com} [copyright holder]

}
\keyword{internal}
