% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_contribution}
\alias{get_contribution}
\title{Calculate Individual Contribution}
\usage{
get_contribution(income, dispoable_allowance)
}
\arguments{
\item{income}{A numeric value expressing the income for this individual.}

\item{dispoable_allowance}{A numeric value expressing the amount set aside as
disposable income.}
}
\value{
A numeric value.
}
\description{
Calculate Individual Contribution
}
\examples{
library(dplyr)

test_data <- tibble::tibble(
  item = c("Item 1", "Item 2", "Item 3", "Item 1", "Item 2"),
  type = c(rep("operating_spend", 3), rep("income", 2)),
  value = c(100,250,50,1000,1200)
)

income <- test_data |>
filter(type == "income" & item == "Item 1") |>
pull(value)

get_contribution(income, 250)
}
