% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_net_income}
\alias{get_net_income}
\title{Calculate Income After Operating Spend}
\usage{
get_net_income(data, period = "monthly")
}
\arguments{
\item{data}{A tibble with three columns: item, type, and value.}

\item{period}{Should the values be calculated as monthly or fortnightly?
Default is monthly.}
}
\value{
A numeric value.
}
\description{
Calculate Income After Operating Spend
}
\examples{
test_data <- tibble::tibble(
  item = c("Item 1", "Item 2", "Item 3", "Item 1", "Item 2"),
  type = c(rep("operating_spend", 3), rep("income", 2)),
  value = c(100,250,50,1000,1200)
)

#Total net income:
get_net_income(test_data)
}
