% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_savings_allowance}
\alias{get_savings_allowance}
\title{Calculate Savings Allowance}
\usage{
get_savings_allowance(
  net_income,
  disposable_income_allowance = 1000,
  period = "monthly",
  value_period = "monthly"
)
}
\arguments{
\item{net_income}{Numeric value that represents total net income from all
sources.}

\item{disposable_income_allowance}{Numeric value that represents disposable
income allowance, default is 1000. Multiplied by 2 in the calculation to
assume that this budget is for a couple.}

\item{period}{Should the values be calculated as monthly or fortnightly?
Default is monthly.}

\item{value_period}{Is this being converted from "monthly" (the default) or
"annually"?}
}
\value{
A numeric value.
}
\description{
Calculate Savings Allowance
}
\examples{
test_data <- tibble::tibble(
  item = c("Item 1", "Item 2", "Item 3", "Item 1", "Item 2"),
  type = c(rep("operating_spend", 3), rep("income", 2)),
  value = c(100,250,50,1000,1200)
)

get_savings_allowance(get_net_income(test_data))
}
