% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_total}
\alias{get_total}
\title{Get Total By Budget Type}
\usage{
get_total(data, type, period = "monthly", value_period = "monthly")
}
\arguments{
\item{data}{A tibble with three columns: item, type, and value.}

\item{type}{The budget type, either "operating_spend" or "income".}

\item{period}{Should the values be calculated as monthly or fortnightly?
Default is monthly.}

\item{value_period}{Is this being converted from "monthly" (the default) or
"annually"?}
}
\value{
A numeric value.
}
\description{
Get Total By Budget Type
}
\examples{
test_data <- tibble::tibble(
  item = c("Item 1", "Item 2", "Item 3", "Item 1", "Item 2"),
  type = c(rep("operating_spend", 3), rep("income", 2)),
  value = c(100,250,50,1000,1200)
)

# Sum operating spend:
get_total(test_data, type = "operating_spend")
}
